fix: autoSave Inconsistency
The Tauri Store plugin has a type mismatch for the autoSave option. 
In TypeScript, it's documented as a boolean, but the Rust backend expects a u64. Setting autoSave to true or false in createStore causes an error:

`
Unhandled Promise Rejection: invalid args `autoSave` for command `create_store`: invalid type: boolean `true`, expected u64
`