feat(dialog) - Support picker mode for open dialog (#3030)
For iOS and Android, there are 2 distinct file picker dialogs, one for
picking media and one for picking documents.

Previously, the filters provided by the user would determing which
picker would be displayed.

However, this has led to undefined behavior when no filter was provided.

To resolve this, we now provide a PickerMode (meant for iOS and
eventually Android) to explicitly define which picker we want to
display.

Eventually, we may need to provide more explicit ways of filtering by
MIME type or having specific modes for ImagePicker or VideoPicker for
ease of use on mobile platforms.

But for now, this is an initial implementation that allows specifying
which UI would be preferable for a file picker on mobile platforms.